refactor: navigation 패키지 추가 및 분리
