+++ app/src/main/java/com/dogandpigs/fitnote/presentation/navigation/BottomNavigationBar.kt
@@ -1,0 +1,62 @@
+package com.dogandpigs.fitnote.presentation.navigation
+
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+
+@Composable
+internal fun BottomNavigationBar(navController: NavHostController) {
+    val items = listOf(
+        NavRoutes.Home,
+        NavRoutes.FirstPage,
+        NavRoutes.SecondPage,
+        NavRoutes.ThirdPage
+    )
+
+    NavigationBar(
+        containerColor = Color.White,
+        contentColor = Color(0xFF3F414E)
+    ) {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+
+        items.forEach { item ->
+            NavigationBarItem(
+                icon = {
+                    Icon(
+                        painter = painterResource(id = item.icon),
+                        contentDescription = stringResource(id = item.titleRes),
+                        modifier = Modifier
+                            .width(26.dp)
+                            .height(26.dp)
+                    )
+                },
+                label = { Text(stringResource(id = item.titleRes), fontSize = 9.sp) },
+                selected = currentRoute == item.route,
+                alwaysShowLabel = false,
+                onClick = {
+                    navController.navigate(item.route) {
+                        navController.graph.startDestinationRoute?.let {
+                            popUpTo(it) { saveState = true }
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+}
